requirements.txt
aiohttp
aiosignal
amqp
asgiref
asttokens
async-timeout
attrs
autobahn
Automat
backcall
billiard
celery
certifi
cffi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
cmake
constantly
contourpy
cryptography
cycler
daphne
decorator
Django
djangorestframework
executing
filelock
fonttools
frozenlist
fsspec
huggingface-hub
hyperlink
idna
incremental
ipdb
ipython
jedi
Jinja2
joblib
kiwisolver
kombu
lit
MarkupSafe
matplotlib
matplotlib-inline
mpmath
multidict
networkx
nltk
numpy
nvidia-cublas-cu11
nvidia-cuda-cupti-cu11
nvidia-cuda-nvrtc-cu11
nvidia-cuda-runtime-cu11
nvidia-cudnn-cu11
nvidia-cufft-cu11
nvidia-curand-cu11
nvidia-cusolver-cu11
nvidia-cusparse-cu11
nvidia-nccl-cu11
nvidia-nvtx-cu11
openai
packaging
pandas
parso
pexpect
pickleshare
Pillow
pip-chill
prompt-toolkit
ptyprocess
pure-eval
pyasn1
pyasn1-modules
pycparser
Pygments
pyOpenSSL
pyparsing
python-dateutil
pytz
PyYAML
redis
regex
requests
safetensors
scikit-learn
scipy
seaborn
sentence-transformers
sentencepiece
service-identity
six
sqlparse
stack-data
sympy
threadpoolctl
tokenizers
tomli
torch
torchvision
tqdm
traitlets
transformers
triton
Twisted
txaio
typing_extensions
tzdata
urllib3
vine
wcwidth
whitenoise
yarl
zope.interface
